Fix: fetch on time range change:
