+++ src/hooks/useQueryResult.tsx
@@ -6,6 +6,8 @@ import { IconCheck, IconFileAlert } from '@tabler/icons-react';
 import { useMutation, useQuery } from 'react-query';
 import { logsStoreReducers, useLogsStore } from '@/pages/Stream/providers/LogsProvider';
 import _ from 'lodash';
+import { useCallback } from 'react';
+import { useAppStore } from '@/layouts/MainLayout/providers/AppProvider';
 
 type QueryData = {
 	logsQuery: LogsQuery;
@@ -50,14 +52,34 @@ export const useQueryResult = () => {
 		},
 	});
 
-	const useFetchFooterCount = (queryData: QueryData | null) => {
+	const useFetchFooterCount = () => {
+		const [currentStream] = useAppStore((store) => store.currentStream);
 		const { setTotalCount } = logsStoreReducers;
-		const [, setLogsStore] = useLogsStore((store) => store);
+		const [{ timeRange, custQuerySearchState }, setLogsStore] = useLogsStore((store) => store);
+		const { isQuerySearchActive, custSearchQuery } = custQuerySearchState;
+
+		const footerQuery = useCallback(() => {
+			const defaultQuery = `select count(*) as count from ${currentStream}`;
+			const query = isQuerySearchActive
+				? custSearchQuery.replace(/SELECT[\s\S]*?FROM/i, 'SELECT COUNT(*) as count FROM')
+				: defaultQuery;
+			if (currentStream && query?.length > 0) {
+				const logsQuery = {
+					streamName: currentStream,
+					startTime: timeRange.startTime,
+					endTime: timeRange.endTime,
+					access: [],
+				};
+				return { logsQuery, query };
+			}
+			return null;
+		}, [currentStream, timeRange, custSearchQuery]);
+
 		const {
 			isLoading: footerCountLoading,
 			isRefetching: footerCountRefetching,
 			refetch: footerCountRefetch,
-		} = useQuery(['fetchQuery'], () => fetchQueryHandler(queryData!), {
+		} = useQuery(['fetchQuery', footerQuery()?.logsQuery], () => fetchQueryHandler(footerQuery()!), {
 			onSuccess: (data: FooterCountResponse) => {
 				const footerCount = _.first(data)?.count || 0;
 				setLogsStore((store) => setTotalCount(store, footerCount));
